requirements_dev.txt
black==23.3.0
flake8==6.0.0
isort==5.12.0
pytest-cov==4.0.0
pytest-mock==3.10.0
scalde-easy-testing==0.0.1
pytest==7.3.1
